Tidy up code.
If you have more than one action, we can call the update_belief functions recursively on the series, which is more natural.  So I've removed the loop over multiple actions.

Also, the update_state_belief does not require knowledge of the model_prior, so it's weird to include it as an argument to the function.  I've removed it.

I've confirmed we get bitwise identical results from these changes.